Add separate script to generate XML files for activities
